Extract GenerateDocTests into build-logic
