+++ build.gradle.kts
@@ -1,6 +1,6 @@
 @file:Suppress("UnstableApiUsage")
 
-import kotlin.io.path.isSymbolicLink
+import com.github.jengelman.gradle.plugins.shadow.buildlogic.GenerateDocTests
 import org.gradle.api.plugins.JavaPlugin.API_ELEMENTS_CONFIGURATION_NAME
 import org.gradle.api.plugins.JavaPlugin.JAVADOC_ELEMENTS_CONFIGURATION_NAME
 import org.gradle.api.plugins.JavaPlugin.RUNTIME_ELEMENTS_CONFIGURATION_NAME
@@ -19,6 +19,7 @@ plugins {
   alias(libs.plugins.pluginPublish)
   alias(libs.plugins.spotless)
   alias(libs.plugins.buildConfig)
+  id("build-logic")
 }
 
 version = providers.gradleProperty("VERSION_NAME").get()
@@ -60,11 +61,11 @@ lint {
 spotless {
   kotlin {
     ktfmt(libs.ktfmt.get().version).googleStyle()
-    target("src/**/*.kt")
+    target("src/**/*.kt", "gradle/build-logic/src/**/*.kt")
   }
   kotlinGradle {
     ktfmt(libs.ktfmt.get().version).googleStyle()
-    target("*.kts")
+    target("*.kts", "gradle/build-logic/*.kts")
   }
 }
 
@@ -263,122 +264,4 @@ tasks.clean {
       // Generated by MkDocs.
       rootDir.resolve("site"),
     )
-}
-
-abstract class GenerateDocTests : DefaultTask() {
-  @get:InputDirectory
-  @get:PathSensitive(PathSensitivity.RELATIVE)
-  abstract val inputDirectory: DirectoryProperty
-
-  @get:OutputDirectory abstract val outputDirectory: DirectoryProperty
-
-  @TaskAction
-  fun generate() {
-    val docRoot = inputDirectory.get().asFile
-
-    val snippets =
-      docRoot
-        .walk()
-        .filter { it.isFile && it.extension == "md" && !it.toPath().isSymbolicLink() }
-        .sortedBy { it.path }
-        .flatMap { file ->
-          val source = file.readText()
-          val relativePath = file.relativeTo(docRoot).invariantSeparatorsPath
-          val parentDir = file.parentFile.relativeTo(docRoot).invariantSeparatorsPath
-          val matcher = pattern.matcher(source)
-          sequence {
-            while (matcher.find()) {
-              var line = 1
-              for (i in 0 until matcher.start().coerceAtMost(source.length)) {
-                if (source[i] == '\n') line++
-              }
-              yield(
-                Snippet(
-                  lang = matcher.group(1),
-                  relativePath = relativePath,
-                  parentDir = parentDir,
-                  lineNumber = line,
-                  snippet = matcher.group(2),
-                  sourceLocation = "${file.toURI()}:$line",
-                )
-              )
-            }
-          }
-        }
-        .toList()
-
-    val groovySnippets = snippets.filter { it.lang == "groovy" }
-    val kotlinSnippets = snippets.filter { it.lang == "kotlin" }
-
-    check(snippets.isNotEmpty()) { "No code snippets found in $docRoot." }
-    check(groovySnippets.size == kotlinSnippets.size) {
-      "All languages must have the same number of code snippets: groovy=${groovySnippets.size}, kotlin=${kotlinSnippets.size}"
-    }
-
-    val outputDir = outputDirectory.get().asFile
-    snippets
-      .groupBy { it.parentDir }
-      .forEach { (parentDir, fileSnippets) ->
-        val className =
-          if (parentDir.isEmpty()) {
-            "ReadmeDocTest"
-          } else {
-            parentDir.split('/', '-').joinToString("") { it.replaceFirstChar(Char::titlecase) } +
-              "DocTest"
-          }
-        val functionsCode =
-          fileSnippets.joinToString("\n\n") {
-            (lang, relativePath, _, lineNumber, snippet, sourceLocation) ->
-            val functionName = "line_${lineNumber}_$lang"
-            """
-            |  @Test
-            |  @DisplayName("$relativePath:$lineNumber ($lang)")
-            |  fun `$functionName`(@TempDir tempDir: Path) {
-            |    SnippetExecutable(
-            |      lang = "$lang",
-            |      snippet =
-            |        ""${'"'}
-            |        $snippet
-            |        ""${'"'},
-            |      sourceLocation = "$sourceLocation",
-            |    ).execute(tempDir)
-            |  }
-            """
-              .trimMargin()
-          }
-        outputDir
-          .resolve("com/github/jengelman/gradle/plugins/shadow/docs/$className.kt")
-          .apply { parentFile.mkdirs() }
-          .writeText(
-            """
-            |package com.github.jengelman.gradle.plugins.shadow.docs
-            |
-            |import com.github.jengelman.gradle.plugins.shadow.SnippetExecutable
-            |import java.nio.file.Path
-            |import org.junit.jupiter.api.DisplayName
-            |import org.junit.jupiter.api.Test
-            |import org.junit.jupiter.api.io.TempDir
-            |
-            |class $className {
-            |$functionsCode
-            |}
-            |
-            """
-              .trimMargin()
-          )
-      }
-  }
-
-  private data class Snippet(
-    val lang: String,
-    val relativePath: String,
-    val parentDir: String,
-    val lineNumber: Int,
-    val snippet: String,
-    val sourceLocation: String,
-  )
-
-  private companion object {
-    val pattern = "(?ims) {4}```(groovy|kotlin)\n(.*?)\n {4}```".toPattern()
-  }
 }
