+++ gradle/build-logic/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+  `kotlin-dsl`
+}
+
+gradlePlugin {
+  plugins {
+    register("build-logic") {
+      implementationClass = "com.github.jengelman.gradle.plugins.shadow.buildlogic.BuildLogicPlugin"
+    }
+  }
+}
