+++ gradle/build-logic/settings.gradle.kts
@@ -1,0 +1,13 @@
+dependencyResolutionManagement {
+  repositories {
+    mavenCentral()
+    gradlePluginPortal()
+  }
+  versionCatalogs {
+    create("libs") {
+      from(files("../libs.versions.toml"))
+    }
+  }
+}
+
+rootProject.name = "build-logic"
